# [AWS] Fix best disk tier on AWS to use the best IOPS